requirements.txt
certifi==2019.9.11
chardet==3.0.4
conda==4.3.16
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.10.3
idna==2.8
pycosat==0.6.3
pydot==1.4.1
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
